Move User message utils to separate file
