+++ api/internal/users/v1/utils.go
@@ -1,0 +1,16 @@
+package v1
+
+import (
+	userspb "buf.build/gen/go/fjarm/fjarm/protocolbuffers/go/fjarm/users/v1"
+	"github.com/fjarm/fjarm/api/internal/logvals"
+)
+
+func redactedUserMessageString(msg *userspb.User) string {
+	if msg == nil {
+		return logvals.Nil
+	}
+	rm := &userspb.User{
+		UserId: msg.UserId,
+	}
+	return rm.String()
+}
